python-server/requirements.txt
fastmcp>=2.12.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
psutil>=5.9.0
